+++ lib/widgets/paste_recipe_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class PastRecipe extends StatelessWidget {
+class PasteRecipe extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
